refactor(download): moves same env upload check to download_configs.go
The reason is that this function is only called in download_configs.go, while being defined in download_command.go. This is feature envy, it doesn't belong to download_command.go, therefore I moved it to where it is used --> download_configs.go
